Make buffers reallocatable
In order to make this possible, all buffers now have to be allocated
indirectly (because a userdata cannot be resized).

In order to make this work, the struct library had to be made aware of
buffers; the simplest way was to merge it into the core library.

All documentation and build system files (including Windows makefiles)
have been updated accordingly.